test(client): rename utils folder
