+++ packages/client/src/__tests__/integration/happy/composites-mongo/__helpers___/build-data/commentOptionalPropDataB.ts
