+++ packages/client/src/__tests__/integration/happy/composites-mongo/findMany/optional.ts
@@ -1,6 +1,6 @@
 import { getTestClient } from '../../../../../utils/getTestClient'
-import { commentOptionalPropDataA } from '../utils/build-data/commentOptionalPropDataA'
-import { commentOptionalPropDataB } from '../utils/build-data/commentOptionalPropDataB'
+import { commentOptionalPropDataA } from '../__helpers___/build-data/commentOptionalPropDataA'
+import { commentOptionalPropDataB } from '../__helpers___/build-data/commentOptionalPropDataB'
 
 const describeIf = (condition: boolean) => (condition ? describe : describe.skip)
